pyproject.toml
[project]
name = "what-if-simulator"
version = "0.1.0"
description = "Local what-if risk score simulator with Databricks data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.6.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "lightgbm>=4.0.0",
  "shap>=0.45.0",
  "mlflow>=2.16.0",
  "databricks-sql-connector>=3.0.0",
]

[project.scripts]
what-if-simulator = "what_if_app.run:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["what_if_app*"]
